Rename collectionTree as collection
