+++ api/api/src/test/scala/uk/ac/wellcome/platform/api/services/CollectionServiceTest.scala
@@ -55,23 +55,23 @@ class CollectionServiceTest
       storeWorks(index)
       whenReady(service.retrieveTree(index, List("a/b"))) { result =>
         result shouldBe Right(
-          CollectionTree(
+          Collection(
             path = CollectionPath("a", CollectionLevel.Collection),
             work = workA,
             children = List(
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/b", CollectionLevel.Series),
                 work = workB,
                 children = List(
-                  CollectionTree(
+                  Collection(
                     path = CollectionPath("a/b/c", CollectionLevel.Item),
                     work = workC),
-                  CollectionTree(
+                  Collection(
                     path = CollectionPath("a/b/d", CollectionLevel.Item),
                     work = workD)
                 )
               ),
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/e", CollectionLevel.Series),
                 work = workE),
             )
@@ -86,31 +86,31 @@ class CollectionServiceTest
       storeWorks(index)
       whenReady(service.retrieveTree(index, List("a/b", "a/e/f"))) { result =>
         result shouldBe Right(
-          CollectionTree(
+          Collection(
             path = CollectionPath("a", CollectionLevel.Collection),
             work = workA,
             children = List(
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/b", CollectionLevel.Series),
                 work = workB,
                 children = List(
-                  CollectionTree(
+                  Collection(
                     path = CollectionPath("a/b/c", CollectionLevel.Item),
                     work = workC),
-                  CollectionTree(
+                  Collection(
                     path = CollectionPath("a/b/d", CollectionLevel.Item),
                     work = workD)
                 )
               ),
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/e", CollectionLevel.Series),
                 work = workE,
                 children = List(
-                  CollectionTree(
+                  Collection(
                     path = CollectionPath("a/e/f", CollectionLevel.Series),
                     work = workF,
                     children = List(
-                      CollectionTree(
+                      Collection(
                         path = CollectionPath("a/e/f/g", CollectionLevel.Item),
                         work = workG)
                     )
@@ -129,15 +129,15 @@ class CollectionServiceTest
       storeWorks(index)
       whenReady(service.retrieveTree(index, List("a"))) { result =>
         result shouldBe Right(
-          CollectionTree(
+          Collection(
             path = CollectionPath("a", CollectionLevel.Collection),
             work = workA,
             children = List(
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/b", CollectionLevel.Series),
                 work = workB,
               ),
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/e", CollectionLevel.Series),
                 work = workE,
               )
@@ -177,11 +177,11 @@ class CollectionServiceTest
       storeWorks(index, List(p, q))
       whenReady(service.retrieveTree(index, List("p/q"))) { result =>
         result shouldBe Right(
-          CollectionTree(
+          Collection(
             path = CollectionPath("p", CollectionLevel.Collection),
             work = p.withData(_.copy(items = Nil)),
             children = List(
-              CollectionTree(
+              Collection(
                 path = CollectionPath("p/q", CollectionLevel.Item),
                 work = q.withData(_.copy(notes = Nil)),
               )
