+++ common/display/src/test/scala/uk/ac/wellcome/display/models/DisplayCollectionTreeTest.scala
@@ -6,7 +6,7 @@ import uk.ac.wellcome.models.work.internal._
 import uk.ac.wellcome.display.models.v2._
 import uk.ac.wellcome.models.work.generators.WorksGenerators
 
-class DisplayCollectionTreeTest
+class DisplayCollectionTest
     extends FunSpec
     with Matchers
     with WorksGenerators {
@@ -19,19 +19,19 @@ class DisplayCollectionTreeTest
     val a = work("a", CollectionLevel.Collection)
     val b = work("a/b", CollectionLevel.Series)
     val c = work("a/b/c", CollectionLevel.Item)
-    val tree = CollectionTree(List(a, b, c)).right.get
-    DisplayCollectionTree(tree, List("a/b/c")) shouldBe
-      DisplayCollectionTree(
+    val tree = Collection(List(a, b, c)).right.get
+    DisplayCollection(tree, List("a/b/c")) shouldBe
+      DisplayCollection(
         path = DisplayCollectionPath("a", "Collection"),
         work = DisplayWorkV2(a),
         children = Some(
           List(
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/b", "Series"),
               work = DisplayWorkV2(b),
               children = Some(
                 List(
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/b/c", "Item"),
                     work = DisplayWorkV2(c),
                     children = Some(Nil)
@@ -52,14 +52,14 @@ class DisplayCollectionTreeTest
           path = "a/b",
           level = CollectionLevel.Item,
           label = Some("!!!"))))
-    val tree = CollectionTree(List(a, b)).right.get
-    DisplayCollectionTree(tree, List("a")) shouldBe
-      DisplayCollectionTree(
+    val tree = Collection(List(a, b)).right.get
+    DisplayCollection(tree, List("a")) shouldBe
+      DisplayCollection(
         path = DisplayCollectionPath("a", "Collection"),
         work = DisplayWorkV2(a),
         children = Some(
           List(
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/b", "Item", Some("!!!")),
               work = DisplayWorkV2(b),
             )
@@ -74,19 +74,19 @@ class DisplayCollectionTreeTest
     val c = work("a/b/c", CollectionLevel.Item)
     val d = work("a/d", CollectionLevel.Series)
     val e = work("a/d/e", CollectionLevel.Item)
-    val tree = CollectionTree(List(a, b, c, d, e)).right.get
-    DisplayCollectionTree(tree, List("a/b/c", "a/d/e")) shouldBe
-      DisplayCollectionTree(
+    val tree = Collection(List(a, b, c, d, e)).right.get
+    DisplayCollection(tree, List("a/b/c", "a/d/e")) shouldBe
+      DisplayCollection(
         path = DisplayCollectionPath("a", "Collection"),
         work = DisplayWorkV2(a),
         children = Some(
           List(
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/b", "Series"),
               work = DisplayWorkV2(b),
               children = Some(
                 List(
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/b/c", "Item"),
                     work = DisplayWorkV2(c),
                     children = Some(Nil)
@@ -94,12 +94,12 @@ class DisplayCollectionTreeTest
                 )
               )
             ),
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/d", "Series"),
               work = DisplayWorkV2(d),
               children = Some(
                 List(
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/d/e", "Item"),
                     work = DisplayWorkV2(e),
                     children = Some(Nil)
@@ -118,24 +118,24 @@ class DisplayCollectionTreeTest
     val c = work("a/b/c", CollectionLevel.Item)
     val d = work("a/d", CollectionLevel.Item)
     val e = work("a/b/e", CollectionLevel.Item)
-    val tree = CollectionTree(List(a, b, c, d, e)).right.get
-    DisplayCollectionTree(tree, List("a/b/c")) shouldBe
-      DisplayCollectionTree(
+    val tree = Collection(List(a, b, c, d, e)).right.get
+    DisplayCollection(tree, List("a/b/c")) shouldBe
+      DisplayCollection(
         path = DisplayCollectionPath("a", "Collection"),
         work = DisplayWorkV2(a),
         children = Some(
           List(
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/b", "Series"),
               work = DisplayWorkV2(b),
               children = Some(
                 List(
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/b/c", "Item"),
                     work = DisplayWorkV2(c),
                     children = Some(Nil)
                   ),
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/b/e", "Item"),
                     work = DisplayWorkV2(e),
                     children = None
@@ -143,7 +143,7 @@ class DisplayCollectionTreeTest
                 )
               )
             ),
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/d", "Item"),
               work = DisplayWorkV2(d),
               children = None
@@ -159,19 +159,19 @@ class DisplayCollectionTreeTest
     val c = work("a/b/c", CollectionLevel.Item)
     val d = work("a/d", CollectionLevel.Series)
     val e = work("a/d/e", CollectionLevel.Item)
-    val tree = CollectionTree(List(a, b, c, d, e)).right.get
-    DisplayCollectionTree(tree, List("a/b")) shouldBe
-      DisplayCollectionTree(
+    val tree = Collection(List(a, b, c, d, e)).right.get
+    DisplayCollection(tree, List("a/b")) shouldBe
+      DisplayCollection(
         path = DisplayCollectionPath("a", "Collection"),
         work = DisplayWorkV2(a),
         children = Some(
           List(
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/b", "Series"),
               work = DisplayWorkV2(b),
               children = Some(
                 List(
-                  DisplayCollectionTree(
+                  DisplayCollection(
                     path = DisplayCollectionPath("a/b/c", "Item"),
                     work = DisplayWorkV2(c),
                     children = None
@@ -179,7 +179,7 @@ class DisplayCollectionTreeTest
                 )
               )
             ),
-            DisplayCollectionTree(
+            DisplayCollection(
               path = DisplayCollectionPath("a/d", "Series"),
               work = DisplayWorkV2(d),
               children = None
