+++ common/internal_model/src/test/scala/uk/ac/wellcome/models/work/internal/CollectionTreeTest.scala
@@ -3,7 +3,7 @@ package uk.ac.wellcome.models.work.internal
 import org.scalatest.{FunSpec, Matchers}
 import uk.ac.wellcome.models.work.generators.WorksGenerators
 
-class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
+class CollectionTest extends FunSpec with Matchers with WorksGenerators {
 
   def work(path: String, level: CollectionLevel) =
     createIdentifiedWorkWith(
@@ -15,26 +15,26 @@ class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
     val c = work("a/b/c", CollectionLevel.Item)
     val d = work("a/d", CollectionLevel.Series)
     val e = work("a/b/e", CollectionLevel.Item)
-    CollectionTree(List(b, d, a, c, e)) shouldBe Right(
-      CollectionTree(
+    Collection(List(b, d, a, c, e)) shouldBe Right(
+      Collection(
         path = CollectionPath("a", CollectionLevel.Collection),
         work = a,
         children = List(
-          CollectionTree(
+          Collection(
             path = CollectionPath("a/b", CollectionLevel.Series),
             work = b,
             children = List(
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/b/c", CollectionLevel.Item),
                 work = c
               ),
-              CollectionTree(
+              Collection(
                 path = CollectionPath("a/b/e", CollectionLevel.Item),
                 work = e
               ),
             ),
           ),
-          CollectionTree(
+          Collection(
             path = CollectionPath("a/d", CollectionLevel.Series),
             work = d
           )
@@ -52,12 +52,12 @@ class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
           level = CollectionLevel.Item,
           label = Some("!!!")))
     )
-    CollectionTree(List(a, b)) shouldBe Right(
-      CollectionTree(
+    Collection(List(a, b)) shouldBe Right(
+      Collection(
         path = CollectionPath("a", CollectionLevel.Collection),
         work = a,
         children = List(
-          CollectionTree(
+          Collection(
             path = CollectionPath("a/b", CollectionLevel.Item, Some("!!!")),
             work = b
           )
@@ -70,7 +70,7 @@ class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
     val x = work("x", CollectionLevel.Collection)
     val y = work("x/y", CollectionLevel.Series)
     val z = work("x/a/z", CollectionLevel.Item)
-    val result = CollectionTree(List(x, y, z))
+    val result = Collection(List(x, y, z))
     result shouldBe a[Left[_, _]]
     result.left.get.getMessage shouldBe "Not all works in collection are connected to root 'x': x/a/z"
   }
@@ -80,13 +80,13 @@ class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
     val y = work("x/y", CollectionLevel.Series)
     val z1 = work("x/y/z", CollectionLevel.Item)
     val z2 = work("x/y/z", CollectionLevel.Item)
-    val result = CollectionTree(List(x, y, z1, z2))
+    val result = Collection(List(x, y, z1, z2))
     result shouldBe a[Left[_, _]]
     result.left.get.getMessage shouldBe "Tree contains duplicate paths: x/y/z"
   }
 
   it("errors creating a tree when empty list") {
-    val result = CollectionTree(Nil)
+    val result = Collection(Nil)
     result shouldBe a[Left[_, _]]
     result.left.get.getMessage shouldBe "Cannot create empty tree"
   }
@@ -95,7 +95,7 @@ class CollectionTreeTest extends FunSpec with Matchers with WorksGenerators {
     val x = work("x", CollectionLevel.Collection)
     val y = work("x/y", CollectionLevel.Series)
     val z = createIdentifiedWorkWith(collectionPath = None)
-    val result = CollectionTree(List(x, y, z))
+    val result = Collection(List(x, y, z))
     result shouldBe a[Left[_, _]]
   }
 }
